requirements.txt
aiogram==3.4.1
openai==1.45.0
python-dotenv==1.0.1
SQLAlchemy==2.0.23
httpx==0.27.0